Add ability for non-versioned attributes
This commit adds the ability to pass in non-versioned attributes to the version history module.
You define the constant in your class, and then we reference that constant when creating
new versions. This constant contains just a list of strings of your attributes, because
that's what they will show up as in the new_version object.